Pretty large refactor of how lighting works in Ardor3D.
Lights are now derived from Spatial, meaning they exist directly in the Scenegraph and can move and rotate accordingly.
LightState is no more.  Lights are gathered by a SceneIndexer and managed by a LightManager in the indexer.  This is where we will get shadow textures back up and running again.
Lights no longer have an ambient or specular color - diffuse is now just "color".  Lights now also have an intensity value that is used during diffuse and specular calcs.  Ambient is now a global setting on LightManager.
The phong light shader is updated to work with the lighting changes.